Added some refactoring
