+++ src/view/editingPoint.js
@@ -1,7 +1,7 @@
 import { TYPES, CITIES } from '../mock/createData.js';
-import { currentTime, createElement } from '../utils/utils.js';
+import { currentTime } from '../utils/utils.js';
 import { addOffers, createTypes, createCities, getFormatTime } from '../utils/renderingUtils.js';
-
+import { createElement } from '../view/renderingUtils.js';
 const editPoint = (points = {}) => {
   const {
     basePrice = 0,
