+++ src/view/newPoint.js
@@ -1,6 +1,7 @@
-import { currentTime, createElement } from '../utils/utils.js';
+import { currentTime } from '../utils/utils.js';
 import {CITIES, TYPES } from '../mock/createData.js';
 import { addOffers, createTypes, createCities, getFormatTime, getPhotos } from '../utils/renderingUtils.js';
+import { createElement } from '../view/renderingUtils.js';
 
 const addNewPoint = (points = {}) => {
   const {
